making a Comment repository and removing createComent() method from ImageRepository into CommentRepository as it is good practice
