+++ src/main/java/ImageHoster/repository/CommentRepository.java
@@ -1,0 +1,39 @@
+package ImageHoster.repository;
+
+import ImageHoster.model.Comment;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.EntityTransaction;
+import javax.persistence.PersistenceUnit;
+
+
+//The annotation is a special type of @Component annotation which describes that the class defines a data repository
+@Repository
+public class CommentRepository {
+
+    //Get an instance of EntityManagerFactory from persistence unit with name as 'imageHoster'
+    @PersistenceUnit(unitName = "imageHoster")
+    private EntityManagerFactory emf;
+
+    //The method receives the comment for a particular image in the database
+    //Creates an instance of EntityManager
+    //Starts a transaction
+    //persist method changes the state of the comment model from transient state to persistent state
+    //The transaction is committed if it is successful
+    //The transaction is rolled back in case of unsuccessful transaction
+    public void  createComment(Comment comment) {
+
+        EntityManager em = emf.createEntityManager();
+        EntityTransaction transaction = em.getTransaction();
+
+        try {
+            transaction.begin();
+            em.persist(comment);
+            transaction.commit();
+        } catch (Exception e) {
+            transaction.rollback();
+        }
+    }
+}
