+++ src/main/java/ImageHoster/service/CommentService.java
@@ -1,7 +1,7 @@
 package ImageHoster.service;
 
 import ImageHoster.model.Comment;
-import ImageHoster.repository.ImageRepository;
+import ImageHoster.repository.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,10 +9,11 @@ import org.springframework.stereotype.Service;
 public class CommentService {
 
     @Autowired
-    private ImageRepository imageRepository;
+    //private ImageRepository imageRepository;
+    private CommentRepository commentRepository;
 
-    //Call the createComment method in the ImageRepository class to persist the comment in the database
+    //Call the createComment method in the CommentRepository class to persist the comment in the database
     public void createComment(Comment comment){
-        imageRepository.createComment(comment);
+        commentRepository.createComment(comment);
     }
 }
